refactor: remove unused date picker components and clean up Playground structure
