--- src/components/core/date-range-picker.tsx
@@ -1,59 +1,0 @@
-"use client";
-
-import * as React from "react";
-import { format } from "date-fns";
-import { CalendarIcon } from "lucide-react";
-
-import { cn } from "@/lib/utils";
-import { Button } from "@/components/ui/button";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
-
-interface IDatePickerWithRangeProps {
-  className?: string;
-  date?: DateRange;
-  setDate?: React.Dispatch<React.SetStateAction<DateRange | undefined>>;
-  placeholder?: string;
-}
-
-export function DatePickerWithRange({
-  className,
-  date,
-  setDate,
-  placeholder = "Pick a date",
-}: React.HTMLAttributes<HTMLDivElement> & IDatePickerWithRangeProps) {
-  return (
-    <div className={cn("grid gap-2", className)}>
-      <Popover>
-        <PopoverTrigger asChild>
-          <Button
-            id="date"
-            variant={"outline"}
-            className={cn(
-              "w-[300px] justify-start text-left font-normal",
-              !date && "text-muted-foreground"
-            )}
-          >
-            <CalendarIcon />
-            {date?.from ? (
-              date.to ? (
-                <>
-                  {format(date.from, "LLL dd, y")} -{" "}
-                  {format(date.to, "LLL dd, y")}
-                </>
-              ) : (
-                format(date.from, "LLL dd, y")
-              )
-            ) : (
-              <span>{placeholder}</span>
-            )}
-          </Button>
-        </PopoverTrigger>
-        <PopoverContent className="w-auto p-0" align="start"></PopoverContent>
-      </Popover>
-    </div>
-  );
-}
